refactor: move feature detection to its own package
